move pytest adoption to root conftest
